basic circle chart created
